Move rest of APIUtils methods to ApiRequest where they are used
